+++ strategic_reports/daily/core/__init__.py
@@ -121,6 +121,7 @@ from .tag_graph import build_graph_data, find_bridge_tags, write_tag_graph
 from .urgency import UrgencyAlert, append_run, check_alerts, load_history
 from .bullet_diff import append_bullet_run, diff_all_topics, load_bullet_history
 from .db import connect as connect_db, ensure_safe_db_path, record_run
+from .article_archive import load_articles, record_articles
 from .tag_tracking import (
     EmergingTagAlert,
     check_emerging_tags,
@@ -164,6 +165,8 @@ __all__ = [
     "connect_db",
     "ensure_safe_db_path",
     "record_run",
+    "load_articles",
+    "record_articles",
     "EmergingTagAlert",
     "check_emerging_tags",
     "load_tag_rate_history",